Support virthost localhost via local connection
Allow local connection transport, when virthost is localhost. Keep
SSH'ing to virthosts, like 127.0.0.2, for backwards compatibility.

We cannot use the local connection transport globally for a full
quickstart run, as this breaks the static inventory generation and all
hosts receive that local transport. To mitigate that, provide a
one-time inventory for the virthost group consisting of localhost only
and using the local connection transport. It is only used for initial
ansible tasks required to build a real hosts in in-memory inventory
(and saved into static inventory later on, which throws that one-time
inventoty away).

Change-Id: Ief903ef2973a7239ee1b2390d4e3a8482d466ea1
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>